Starting cookbook with AWS-Amplify from scratch after hitting a roadblock with cookbook-with-cdk
Details:
cookbook-with-cdk doesn't allow me to mock the AmplifyAuthentication for unit testing (with jest)
since I created the AppSync using CDK instead of through Amplify.
It looks like AWS has a lot of issues when trying to mix and match different technologies.
Therefore, I've decided to do everything using Amplify to avoid this types of issues.